+++ app/src/main/java/com/luna/budgetapp/data/local/entity/ExpensePresetEntity.kt
@@ -9,7 +9,7 @@ import java.time.LocalDateTime
 data class ExpensePresetEntity(
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "remote_id") val remoteId: String? = null,
-    @ColumnInfo(name = "amount") val amount: Double,
+    @ColumnInfo(name = "amount") val amount: Long,
     @ColumnInfo(name = "category") val category: String,
     @ColumnInfo(name = "type") val type: String,
     @ColumnInfo(name = "created_at") val createdAt: LocalDateTime
