+++ app/src/main/java/com/luna/budgetapp/data/remote/dto/ExpenseDto.kt
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 data class ExpenseDto(
     @SerializedName("id") val id: Long,
     @SerializedName("name") val name: String?,
-    @SerializedName("amount") val amount: Double,
+    @SerializedName("amount") val amount: Long,
     @SerializedName("category") val category: String,
     @SerializedName("type") val type: String,
     @SerializedName("date") val date: LocalDateTime
