+++ app/src/main/java/com/luna/budgetapp/presentation/screen/expenselist/ExpenseListStateEvents.kt
@@ -18,7 +18,7 @@ sealed interface UiState {
 }
 
 data class ExpensesState(
-    val totalAmount: Double = 0.0
+    val totalAmount: Long = 0L
 )
 
 data class DateState(
